Add certificate validation
This change adds support for a certificate trust store. When
performing signature verification, all certificates in the trust
store are loaded into a certificate verification context. This
context is used to validate the signing certificate, verifying
that the certificate belongs to a valid certificate chain rooted
in the trust store.

The signature_utils.get_verifier function is updated to accept
an additional, optional parameter: trust_store_path. This
parameter should contain a valid filesystem path to the
directory acting as the certificate trust store. If not
provided, it defaults to None and the trust store will be
considered empty.

All new certificate utility code is added in a new module named
certificate_utils.

For more information on this work, see the spec:
https://review.openstack.org/#/c/357151/

SecurityImpact
DocImpact

Change-Id: I8d7f43fb4c0573ac3681147eac213b369bbbcb3b