Address verifier DeprecationWarning
The use of signer and verifier in cryptography has been
deprecated, and causes the following warning:

cursive/cursive/signature_utils.py:139: DeprecationWarning: signer
and verifier have been deprecated. Please use sign and verify
instead.

This patch adds a wrapper around the use of verifier, so
that sign and verify are used with cryptography, but the
client use of the library doesn't have to change.

Change-Id: Ib4aaa4fc9eb893b74f08bc8ff732a4dae152f685